accessing_apis/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
praw==7.1.0
prawcore==1.5.0
psaw==0.0.12
pymongo==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
